Final commit

+++ ProjectPlanner.API/Dtos/ProjectForListDto.cs
@@ -15,7 +15,7 @@ namespace ProjectPlanner.API.Dtos
         public DateTime Created { get; set; }
         public DateTime Modified { get; set; }
         public DateTime EstimatedDate { get; set; }
-        public int  CompletedPercentage { get; set; }
+        public double  CompletedPercentage { get; set; }
         public string OwnerId { get; set; }
         public UserForListDto Owner { get; set; }
         public ICollection<UserForListDto> Collaborators { get; set; }
